# routes/api.yml
openapi: 3.1.0
info:
  title: Code Golf API
  description: API for [Code Golf](https://code.golf).
  version: unstable
servers:
  - url: https://code.golf/api
...
paths:
...
  /golfers/{name}:
    get:
      operationId: getGolfersByName
      parameters:
        - name: name
          description: Golfer Name
          in: path
          required: true
          schema:
            type: string
      tags: [golfers]
      summary: Get a golfer by name
      responses:
        '200':
          description: A single golfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Golfer'
...
  /suggestions/golfers:
    get:
      operationId: suggestGolfers
      parameters:
        - name: q
          description: Case-insensitive substring to search for
          in: query
          schema:
            type: string
      tags: [suggestions]
      summary: Returns a list of matching golfers, limited to 10
      responses:
        '200':
          description: A JSON array matching golfer names, limited to 10
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-session
  schemas:
...
    Golfer:
      type: object
      properties:
        admin:
          type: boolean
        country:
          type:
            - string
            - 'null'
        id:
          type: integer
        name:
          type: string
        pronouns:
          type:
            - string
            - 'null'
        sponsor:
          type: boolean
        started:
          type: string
          format: date-time
...